requirements.txt
Flask==2.0.1
plaid-python==8.1.0
pandas==1.3.2
scikit-learn==0.24.2
numpy==1.21.2
Flask-Cors==3.0.10
